Create collection javascript file
